Add mock module for queue APIs
